requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==3.0.4
channels-redis==4.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
distlib==0.3.9
Django==4.2.9
django-bootstrap4==24.4
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==36.1.1
filelock==3.16.1
frozenlist==1.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
pbr==6.1.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg==3.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-django==4.10.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
service-identity==24.2.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stevedore==5.4.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.1
virtualenv-clone==0.5.7
wcwidth==0.2.13
yarl==1.18.3
zope.interface==7.2
